screen-orientation/event-before-promise.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script>
  promise_test(async t => {
    await test_driver.bless("request full screen", () => {
      return document.documentElement.requestFullscreen();
    });

    const promiseToChange = new Promise(resolve => {
      screen.orientation.addEventListener("change", resolve);
    });

    function firstToComplete(response) {
      if (response instanceof Event) {
        return "event";
      } else if (response === undefined) {
        return "promise";
      }
      throw new Error("unexpected result!");
    }

    async function orderCheck() {
      try {
        const result = await Promise.race([
          screen.orientation.lock("landscape"),
          promiseToChange
        ]);
        return firstToComplete(result);
      } catch (err) {
        console.log(`error: ${err}`);
      }
    }
    const orderResult = await orderCheck();
    assert_equals(orderResult, "event");
  }, "Event is fired before promise is resolved");
</script>
